backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-chess>=1.10.0
numpy>=1.26.0
google-auth>=2.40.0
websockets>=12.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
